[TVMC] Split common tvmc file into more specific files
This follows from #9206 and splits common.py into multiple smaller and more focussed files.
