+++ python/tvm/driver/tvmc/__init__.py
@@ -14,10 +14,19 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint: disable=redefined-builtin
+# pylint: disable=redefined-builtin,wrong-import-position
 """
 TVMC - TVM driver command-line interface
 """
+
+
+class TVMCException(Exception):
+    """TVMC Exception"""
+
+
+class TVMCImportError(TVMCException):
+    """TVMC TVMCImportError"""
+
 
 from . import micro
 from . import runner
