+++ python/tvm/driver/tvmc/autotuner.py
@@ -34,11 +34,12 @@ from tvm.autotvm.tuner import RandomTuner
 from tvm.autotvm.tuner import XGBTuner
 from tvm.target import Target
 
-from . import common, composite_target, frontends
-from .common import TVMCException
+from . import TVMCException, composite_target, frontends
 from .main import register_parser
 from .model import TVMCModel
-from .target import generate_target_args, reconstruct_target_args
+from .target import target_from_cli, generate_target_args, reconstruct_target_args
+from .shape_parser import parse_shape_string
+from .transform import convert_graph_layout
 
 
 # pylint: disable=invalid-name
@@ -220,7 +221,7 @@ def add_tune_parser(subparsers, _):
         "--input-shapes",
         help="specify non-generic shapes for model to run, format is "
         '"input_name:[dim1,dim2,...,dimn] input_name2:[dim1,dim2]"',
-        type=common.parse_shape_string,
+        type=parse_shape_string,
     )
 
 
@@ -256,9 +257,7 @@ def drive_tune(args):
         logger.info("RPC tracker port: %s", rpc_port)
 
         if not args.rpc_key:
-            raise common.TVMCException(
-                "need to provide an RPC tracker key (--rpc-key) for remote tuning"
-            )
+            raise TVMCException("need to provide an RPC tracker key (--rpc-key) for remote tuning")
     else:
         rpc_hostname = None
         rpc_port = None
@@ -376,7 +375,7 @@ def tune_model(
     tuning_records : str
         The path to the produced tuning log file.
     """
-    target, extra_targets = common.target_from_cli(target, additional_target_options)
+    target, extra_targets = target_from_cli(target, additional_target_options)
     target, target_host = Target.check_and_update_host_consist(target, target_host)
     # TODO(jwfromm) Remove this deepcopy once AlterOpLayout bug that mutates source
     # model is fixed. For now, creating a clone avoids the issue.
@@ -399,7 +398,7 @@ def tune_model(
 
     if rpc_key:
         if hostname is None or port is None:
-            raise common.TVMCException(
+            raise TVMCException(
                 "You must provide a hostname and port to connect to a remote RPC device."
             )
         if isinstance(port, str):
@@ -520,7 +519,7 @@ def autotvm_get_tuning_tasks(
     target, target_host = Target.check_and_update_host_consist(target, target_host)
 
     if alter_layout:
-        mod = common.convert_graph_layout(mod, alter_layout)
+        mod = convert_graph_layout(mod, alter_layout)
 
     tasks = autotvm.task.extract_from_program(
         mod["main"],
@@ -569,7 +568,7 @@ def autoscheduler_get_tuning_tasks(
     target, target_host = Target.check_and_update_host_consist(target, target_host)
 
     if alter_layout:
-        mod = common.convert_graph_layout(mod, alter_layout)
+        mod = convert_graph_layout(mod, alter_layout)
 
     # Extract the tasks
     tasks, task_weights = auto_scheduler.extract_tasks(
